+++ automl_alex/models/model_xgboost.py
@@ -102,7 +102,7 @@ class XGBoost(ModelBase):
             X (pd.DataFrame, shape (n_samples, n_features)): the input data
             y (pd.DataFrame, shape (n_samples, ) or (n_samples, n_outputs)): the target data
         Return:
-            None
+            self
         """
         params = self.model_param.copy()