+++ tests/test_optimizer.py
@@ -16,37 +16,33 @@ RANDOM_SEED = 42
 # ### Test #####################################################
 # Binary-Classification
 
-def test_model_opt_default_classification():
+def test_bestsinglemodelclassifier_default_classification():
     for data_id in [179,4135,]:
         dataset = fetch_openml(data_id=data_id, as_frame=True)
         dataset.target = dataset.target.astype('category').cat.codes
-        X_train, X_test, y_train, y_test = train_test_split(dataset.data, 
-                                                            dataset.target,
+        X_train, X_test, y_train, y_test = train_test_split(dataset.data[:1000], 
+                                                            dataset.target[:1000],
                                                             test_size=0.2, 
                                                             random_state=RANDOM_SEED,)
         de = DataPrepare(normalization=True,verbose=0)
         X_train = de.fit_transform(X_train)
         X_test = de.transform(X_test)
-        for model_name in all_models.keys():
-            print(model_name)
-            model = all_models[model_name](type_of_estimator='classifier',
-                                            random_state=RANDOM_SEED)
 
-            optimizer = automl_alex.BestSingleModelClassifier(models_names = ['LinearModel','KNeighbors','LightGBM','ExtraTrees','XGBoost'],)
-            history = optimizer.opt(X_train, y_train, timeout=300, verbose=3)
-            predicts = optimizer.predict(X_test)
+        optimizer = automl_alex.BestSingleModelClassifier(models_names = ['LinearModel','KNeighbors','RandomForest','LightGBM','ExtraTrees','MLP'])
+        history = optimizer.opt(X_train, y_train, timeout=400, verbose=3)
+        predicts = optimizer.predict(X_test)
 
-            score = round(sklearn.metrics.roc_auc_score(y_test, predicts),4)
-            assert score is not None
-            assert 0.6 < score <= 1
+        score = round(sklearn.metrics.roc_auc_score(y_test, predicts),4)
+        assert score is not None
+        assert 0.5 < score <= 1
 
 
 def test_optimizer_default_classification():
     for data_id in [179,4135,]:
         dataset = fetch_openml(data_id=data_id, as_frame=True)
         dataset.target = dataset.target.astype('category').cat.codes
-        X_train, X_test, y_train, y_test = train_test_split(dataset.data, 
-                                                            dataset.target,
+        X_train, X_test, y_train, y_test = train_test_split(dataset.data[:1000], 
+                                                            dataset.target[:1000],
                                                             test_size=0.2, 
                                                             random_state=RANDOM_SEED,)
         de = DataPrepare(normalization=True,verbose=0)
@@ -58,7 +54,7 @@ def test_optimizer_default_classification():
                                                 random_state=RANDOM_SEED)
             #model = model.fit(X_train, y_train)
 
-            history = model.opt(X_train, y_train, timeout=300, verbose=3)
+            history = model.opt(X_train, y_train, timeout=100, verbose=3)
             predicts = model.predict(X_test)
 
             score = round(sklearn.metrics.roc_auc_score(y_test, predicts),4)