+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.4]
+### ADD
+- Target Encoding in CrossValidation
+- DenoisingAutoencoder in DataPrepare
+- Docs
 
 ## [1.3.1]
 ### Fix